refactor: move extraction engine into dedicated package
