+++ docs/property_extraction.md
@@ -1,6 +1,6 @@
 # Estrazione proprietà testuali
 
-Il modulo `features/extractors.py` implementa un motore di estrazione basato su pattern regex configurabili. Ogni `Pattern` associa un `property_id` a una lista di espressioni regolari e a una sequenza di normalizzatori dichiarati nel pacchetto di estrattori (`extractors_pack`). I normalizzatori built-in spaziano da conversioni numeriche (`to_float`, `to_int`) a manipolazioni di stringa (`lower`, `strip`) e trasformazioni dominio-specifiche (`EI_from_any`, `normalize_foratura`).
+Il pacchetto `robimb.extraction` implementa un motore di estrazione basato su pattern regex configurabili. Ogni `Pattern` associa un `property_id` a una lista di espressioni regolari e a una sequenza di normalizzatori dichiarati nel pacchetto di estrattori (`extractors_pack`). I normalizzatori built-in spaziano da conversioni numeriche (`to_float`, `to_int`) a manipolazioni di stringa (`lower`, `strip`) e trasformazioni dominio-specifiche (`EI_from_any`, `normalize_foratura`). Il modulo storico `features/extractors.py` resta come livello di compatibilità che re-esporta le nuove API.
 
 La funzione `_compile_patterns` filtra i pattern in base all'elenco di proprietà consentite e produce oggetti immutabili efficienti per la fase di matching. Durante l'estrazione (`extract_properties`) ogni regex viene applicata al testo, i match vengono normalizzati e aggregati secondo le regole definite; l'opzione `collect_many` consente di mantenere liste di valori anziché singoli campi.
