+++ src/robimb/extraction/__init__.py
@@ -1,0 +1,18 @@
+"""Property extraction package."""
+
+from .dsl import ExtractorsPack, PatternSpec, PatternSpecs
+from .engine import Pattern, dry_run, extract_properties
+from .normalizers import BUILTIN_NORMALIZERS, Normalizer, NormalizerFactory, build_normalizer
+
+__all__ = [
+    "ExtractorsPack",
+    "PatternSpec",
+    "PatternSpecs",
+    "Pattern",
+    "dry_run",
+    "extract_properties",
+    "BUILTIN_NORMALIZERS",
+    "Normalizer",
+    "NormalizerFactory",
+    "build_normalizer",
+]
