+++ src/robimb/features/extractors.py
@@ -1,383 +1,31 @@
+"""Compatibility layer exposing the extraction engine under ``robimb.features``."""
 
 from __future__ import annotations
-import re
-from dataclasses import dataclass
-from typing import Any, Callable, Dict, Iterable, List, Optional, Sequence, Set, Tuple
 
-# ---------- Normalizers registry ----------
+from ..extraction import (
+    BUILTIN_NORMALIZERS,
+    Normalizer,
+    NormalizerFactory,
+    Pattern,
+    build_normalizer,
+    dry_run,
+    extract_properties,
+)
+from ..extraction.dsl import ExtractorsPack, PatternSpec, PatternSpecs
 
-NormalizerFn = Callable[[Any, str], Any]
-# value, matched_text -> new_value
+# Backwards compatible alias used by older code.
+NormalizerFn = NormalizerFactory
 
-def _comma_to_dot(v, m): 
-    if isinstance(v, str): return v.replace(",", ".")
-    if isinstance(v, list): return [str(x).replace(",", ".") for x in v]
-    return v
-
-def _dot_to_comma(v, m):
-    if isinstance(v, str): return v.replace(".", ",")
-    if isinstance(v, list): return [str(x).replace(".", ",") for x in v]
-    return v
-
-def _to_float(v, m):
-    def conv(x):
-        try: return float(str(x).replace(",", "."))
-        except: return x
-    if isinstance(v, list): return [conv(x) for x in v]
-    return conv(v)
-
-def _to_int(v, m):
-    def conv(x):
-        try: return int(re.sub(r"[^0-9-]", "", str(x)))
-        except: return x
-    if isinstance(v, list): return [conv(x) for x in v]
-    return conv(v)
-
-def _lower(v, m): return v.lower() if isinstance(v, str) else v
-def _upper(v, m): return v.upper() if isinstance(v, str) else v
-def _strip(v, m): return v.strip() if isinstance(v, str) else v
-
-def _ei_from_any(v, m):
-    s = " ".join(v) if isinstance(v, (list, tuple)) else str(v)
-    mnum = re.search(r"(15|30|45|60|90|120|180|240)", s)
-    return f"EI {mnum.group(1)}" if mnum else s
-
-def _dims_join(v, m):
-    # accept tuple/list like ('60','60') -> '60×60'
-    if isinstance(v, (list, tuple)):
-        if len(v) == 2: return f"{v[0]}×{v[1]}"
-        # list of pairs
-        if len(v) > 2 and all(isinstance(x, (list, tuple)) and len(x)==2 for x in v):
-            return "; ".join([f"{a}×{b}" for a,b in v])
-    return v
-
-def _collect_many(v, m):
-    # if list, keep list; else wrap into list
-    if isinstance(v, list): return v
-    return [v]
-
-def _if_cm_to_mm(v, m):
-    # multiply by 10 if match string contains 'cm' (not mm)
-    def conv_one(x):
-        try:
-            val = float(str(x).replace(",", "."))
-        except:
-            return x
-        s = m.lower()
-        if "cm" in s and "mm" not in s:
-            return round(val * 10.0, 3)
-        return val
-    if isinstance(v, list): return [conv_one(x) for x in v]
-    return conv_one(v)
-
-def _unique_list(v, m):
-    if not isinstance(v, list): return v
-    seen=set(); out=[]
-    for x in v:
-        if x not in seen:
-            out.append(x); seen.add(x)
-    return out
-
-def _as_string(v, m):
-    if isinstance(v, list):
-        return [str(x) for x in v]
-    return str(v)
-
-def _concat_dims(v, m):
-    if isinstance(v, (list, tuple)) and len(v) == 2:
-        return f"{v[0]}×{v[1]}"
-    return v
-
-_EI_PATTERN = re.compile(r"(15|30|45|60|90|120|180|240)")
-
-def _format_ei_from_last_int(v, m):
-    if isinstance(v, (list, tuple)):
-        hay = " ".join(str(x) for x in v)
-    else:
-        hay = str(v)
-    nums = _EI_PATTERN.findall(hay)
-    if not nums:
-        nums = _EI_PATTERN.findall(m)
-    if not nums:
-        return v
-    return f"EI {nums[-1]}"
-
-def _take_last_int_to_ei(v, m):
-    return _format_ei_from_last_int(v, m)
-
-
-def _to_ei_class(v, m):
-    formatted = _format_ei_from_last_int(v, m)
-    if isinstance(formatted, str):
-        return formatted.replace(" ", "").upper()
-    return formatted
-
-_FORATURA_MAP = {
-    "pieno": "pieno",
-    "forato": "forato",
-    "semi pieno": "semipieno",
-    "semi-pieno": "semipieno",
-    "semipieno": "semipieno",
-}
-
-def _normalize_foratura(v, m):
-    def norm_one(x):
-        key = str(x).strip().lower().replace("  ", " ")
-        return _FORATURA_MAP.get(key, x)
-    if isinstance(v, list):
-        return [norm_one(x) for x in v]
-    return norm_one(v)
-
-def _cm_to_mm_optional(v, m):
-    lower = m.lower()
-    convert = "cm" in lower and "mm" not in lower
-
-    def conv_one(x):
-        if not convert:
-            return x
-        try:
-            return float(x) * 10.0
-        except Exception:
-            try:
-                return float(str(x).replace(",", ".")) * 10.0
-            except Exception:
-                return x
-
-    if isinstance(v, list):
-        return [conv_one(x) for x in v]
-    return conv_one(v)
-
-
-def _cm_to_mm_if_cm(v, m):
-    def _to_number(x):
-        if isinstance(x, (int, float)):
-            return float(x)
-        try:
-            return float(str(x).replace(",", "."))
-        except Exception:
-            return x
-
-    if isinstance(v, (list, tuple)) and v:
-        value = v[0]
-        unit = v[1] if len(v) > 1 else ""
-        num = _to_number(value)
-        if isinstance(num, (int, float)):
-            unit_norm = str(unit).strip().lower()
-            if unit_norm == "cm":
-                return num * 10.0
-            if unit_norm == "mm":
-                return num
-        return v
-
-    num = _to_number(v)
-    return num
-
-
-def _map_enum_factory(mapping: Dict[str, str]):
-    def _map_enum(v, m):
-        def map_one(x):
-            key = str(x).strip().lower()
-            return mapping.get(key, x)
-        if isinstance(v, list): return [map_one(x) for x in v]
-        return map_one(v)
-    return _map_enum
-
-
-_STRATI_WORDS = {
-    "doppia": 2,
-    "tripla": 3,
-}
-
-
-def _to_strati_count(v, m):
-    if isinstance(v, (list, tuple)) and v:
-        base = str(v[-1])
-    else:
-        base = str(v)
-    key = base.strip().lower()
-    if key in _STRATI_WORDS:
-        return _STRATI_WORDS[key]
-    try:
-        return int(re.sub(r"[^0-9]", "", key))
-    except Exception:
-        digits = re.findall(r"\d+", m)
-        if digits:
-            try:
-                return int(digits[-1])
-            except Exception:
-                pass
-    return v
-
-
-_TIPO_LASTRA_MAP = {
-    "gkb": "standard",
-    "standard": "standard",
-    "gkfi": "fuoco",
-    "gkf": "fuoco",
-    "gklo": "fuoco",
-    "fuoco": "fuoco",
-    "hf": "fuoco",
-    "h2": "idr",
-    "idro": "idr",
-    "idr": "idr",
-    "idrorepellente": "idr",
-    "acustico": "acustica",
-    "acustica": "acustica",
-    "antimuffa": "antimuffa",
-    "fibrocemento": "fibrocemento",
-}
-
-
-def _map_tipo_lastra_enum(v, m):
-    def infer_from_text(text: str) -> Optional[str]:
-        lower = text.strip().lower()
-        for key, value in _TIPO_LASTRA_MAP.items():
-            if key in lower:
-                return value
-        return None
-
-    if isinstance(v, list):
-        return [_map_tipo_lastra_enum(x, m) for x in v]
-
-    match_hint = infer_from_text(m)
-    value_hint = infer_from_text(str(v))
-
-    if value_hint is not None:
-        return value_hint
-    if match_hint is not None:
-        return match_hint
-    return v
-
-BUILTINS: Dict[str, NormalizerFn] = {
-    "comma_to_dot": _comma_to_dot,
-    "dot_to_comma": _dot_to_comma,
-    "to_float": _to_float,
-    "to_int": _to_int,
-    "lower": _lower,
-    "upper": _upper,
-    "strip": _strip,
-    "EI_from_any": _ei_from_any,
-    "dims_join": _dims_join,
-    "collect_many": _collect_many,
-    "if_cm_to_mm": _if_cm_to_mm,
-    "unique_list": _unique_list,
-    "as_string": _as_string,
-    "concat_dims": _concat_dims,
-    "format_EI_from_last_int": _format_ei_from_last_int,
-    "take_last_int->EI {n}": _take_last_int_to_ei,
-    "normalize_foratura": _normalize_foratura,
-    "cm_to_mm?": _cm_to_mm_optional,
-    "to_ei_class": _to_ei_class,
-    "cm_to_mm_if_cm": _cm_to_mm_if_cm,
-    "to_strati_count": _to_strati_count,
-    "map_tipo_lastra_enum": _map_tipo_lastra_enum,
-    # dynamic "map_enum:<name>" supported below
-}
-
-@dataclass
-class Pattern:
-    property_id: str
-    regex: List[str]
-    normalizers: List[str]
-
-def _compile_patterns(extractors_pack: Dict, allowed_properties: Optional[Iterable[str]] = None) -> List[Pattern]:
-    pats: List[Pattern] = []
-    allowed: Optional[Set[str]] = None
-    if allowed_properties is not None:
-        allowed = {p for p in allowed_properties if p}
-    for item in extractors_pack.get("patterns", []):
-        pid = item["property_id"]
-        if allowed is not None and pid not in allowed:
-            continue
-        pats.append(Pattern(
-            property_id=pid,
-            regex=item.get("regex", []),
-            normalizers=item.get("normalizers", [])
-        ))
-    return pats
-
-def _build_normalizer(name: str, extractors_pack: Dict) -> NormalizerFn:
-    # allow "map_enum:<key>" to map values using pack.normalizers[<key>]
-    if name.startswith("map_enum:"):
-        key = name.split(":",1)[1]
-        mapping = extractors_pack.get("normalizers", {}).get(key, {})
-        return _map_enum_factory(mapping)
-    fn = BUILTINS.get(name)
-    if fn is None:
-        # no-op unknown
-        return lambda v,m: v
-    return fn
-
-def _apply_normalizers(value: Any, matched_text: str, normalizers: Sequence[str], extractors_pack: Dict) -> Any:
-    v = value
-    for n in normalizers:
-        fn = _build_normalizer(n, extractors_pack)
-        v = fn(v, matched_text)
-    return v
-
-def _coerce_capture(match: re.Match) -> Any:
-    # No groups: return entire matched text
-    if match.lastindex is None:
-        return match.group(0)
-    # 1 group: return that group
-    if match.lastindex == 1:
-        return match.group(1)
-    # 2+ groups: return tuple of groups
-    return tuple(match.group(i) for i in range(1, match.lastindex+1))
-
-def extract_properties(text: str, extractors_pack: Dict, allowed_properties: Optional[Iterable[str]] = None) -> Dict[str, Any]:
-    """
-    Applica i pattern in ordine; per default prende il primo match per proprietà.
-    Se presente 'collect_many' nei normalizers, accumula tutti i match (deduplicabili con 'unique_list').
-    'dims_join' trasforma tuple di due catture in 'AxB'.
-    'if_cm_to_mm' converte cm→mm se nel match compare 'cm'.
-    'map_enum:<name>' mappa valori verso enum/tabellati definiti nel pack.
-    """
-    out: Dict[str, Any] = {}
-    pats = _compile_patterns(extractors_pack, allowed_properties=allowed_properties)
-    # precompile all regex for speed
-    compiled = [(p, [re.compile(rx, flags=re.IGNORECASE) for rx in p.regex]) for p in pats]
-    for pat, regs in compiled:
-        has_collect = "collect_many" in pat.normalizers
-        for rx in regs:
-            for m in rx.finditer(text):
-                cap = _coerce_capture(m)
-                # normalize immediate dims (tuples) before further steps
-                val = cap
-                # if dims_join later, keep as tuple until normalization
-                val = _apply_normalizers(val, m.group(0), pat.normalizers, extractors_pack)
-                if has_collect:
-                    prev = out.get(pat.property_id, [])
-                    if not isinstance(prev, list):
-                        prev = [prev]
-                    prev.append(val)
-                    out[pat.property_id] = prev
-                else:
-                    # only first capture if not set yet
-                    if pat.property_id not in out:
-                        out[pat.property_id] = val
-            # if property found and not collecting, skip remaining regexes
-            if pat.property_id in out and not has_collect:
-                break
-        # post-process unique_list if needed
-        if has_collect and pat.property_id in out and "unique_list" in pat.normalizers:
-            out[pat.property_id] = BUILTINS["unique_list"](out[pat.property_id], "")
-    return out
-
-def dry_run(text: str, extractors_pack: Dict, allowed_properties: Optional[Iterable[str]] = None) -> Dict[str, Any]:
-    """
-    Restituisce dettagli di matching per debug.
-    """
-    details = []
-    pats = _compile_patterns(extractors_pack, allowed_properties=allowed_properties)
-    for pat in pats:
-        for rx in pat.regex:
-            comp = re.compile(rx, re.IGNORECASE)
-            for m in comp.finditer(text):
-                details.append({
-                    "property_id": pat.property_id,
-                    "regex": rx,
-                    "match": m.group(0),
-                    "groups": m.groups()
-                })
-    return {"matches": details, "extracted": extract_properties(text, extractors_pack, allowed_properties=allowed_properties)}
+__all__ = [
+    "Normalizer",
+    "NormalizerFactory",
+    "NormalizerFn",
+    "Pattern",
+    "PatternSpec",
+    "PatternSpecs",
+    "ExtractorsPack",
+    "BUILTIN_NORMALIZERS",
+    "build_normalizer",
+    "extract_properties",
+    "dry_run",
+]
